+++ CHANGELOG.md
@@ -26,6 +26,7 @@ The following changes have been completed, and will be included in the next tagg
 ### New
 
 * Middleware
+  - Added `AsyncMiddleware` function signature
   - Added `Middleware` function signature
   - Added `MiddlewareStack` value type
 * Errors